+++ js/main.js
@@ -1,1 +1,5 @@
-import './post-renderer'
+import { renderPost } from './render-post.js'
+
+const POST_AMOUNT = 25;
+
+renderPost(POST_AMOUNT);
