+++ js/render-post.js
@@ -1,0 +1,5 @@
+import { generatePost } from './generate-post.js'
+
+const renderPost = (total) => new Array(total).fill(null).map(() => generatePost(total));
+
+export { renderPost }
